# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MenuItem {
  id: string;
  name: string;
  description: string;
  price: number;
  image: string;
  category: string;
  rating: number;
  isBestSeller: boolean;
  isAvailable: boolean;
  dailyQuota: number;
  soldToday: number;
  quotaResetDate: string;
  createdAt: string;
}

export interface MenuItemInput {
  id: string;
  name: string;
  description?: string;
  price: number;
  image?: string;
  category?: string;
  rating?: number;
  isBestSeller?: boolean;
  isAvailable?: boolean;
  dailyQuota?: number;
  soldToday?: number;
  quotaResetDate?: string;
}

export interface MenuItemUpdate {
  name?: string;
  description?: string;
  price?: number;
  image?: string;
  category?: string;
  rating?: number;
  isBestSeller?: boolean;
  isAvailable?: boolean;
  dailyQuota?: number;
  soldToday?: number;
  quotaResetDate?: string;
}

export type MenuItemSoldUpdateItemsItem = {
  id: string;
  quantity: number;
};

export interface MenuItemSoldUpdate {
  items: MenuItemSoldUpdateItemsItem[];
}

export type OrderItems = { [key: string]: unknown };

export interface Order {
  id: string;
  items: OrderItems;
  customerName: string;
  whatsapp: string;
  address: string;
  notes?: string;
  paymentMethod: string;
  status: string;
  totalPrice: number;
  subtotal: number;
  discount: number;
  couponCode?: string | null;
  buktiTransfer?: string | null;
  createdAt: string;
}

export type OrderInputItems = { [key: string]: unknown };

export interface OrderInput {
  id: string;
  items: OrderInputItems;
  customerName: string;
  whatsapp: string;
  address: string;
  notes?: string;
  paymentMethod: string;
  status?: string;
  totalPrice: number;
  subtotal: number;
  discount?: number;
  couponCode?: string | null;
}

export interface OrderUpdate {
  status?: string;
  buktiTransfer?: string | null;
}

export interface StoreSettings {
  storeName: string;
  whatsappNumber: string;
  bankName: string;
  accountNumber: string;
  logoUrl: string;
  bannerUrl: string;
  address: string;
  openTime: string;
  closeTime: string;
  adminUsername: string;
}

export interface StoreSettingsInput {
  storeName?: string;
  whatsappNumber?: string;
  bankName?: string;
  accountNumber?: string;
  logoUrl?: string;
  bannerUrl?: string;
  address?: string;
  openTime?: string;
  closeTime?: string;
  adminUsername?: string;
  adminPasswordHash?: string;
}

export interface Banner {
  id: string;
  title: string;
  subtitle: string;
  image: string;
  bgColor: string;
  sortOrder: number;
  createdAt: string;
}

export interface BannerInput {
  id: string;
  title: string;
  subtitle?: string;
  image?: string;
  bgColor?: string;
  sortOrder?: number;
}

export interface Coupon {
  id: string;
  code: string;
  type: string;
  value: number;
  minOrder: number;
  isActive: boolean;
  createdAt: string;
}

export interface CouponInput {
  id: string;
  code: string;
  type: string;
  value: number;
  minOrder?: number;
  isActive?: boolean;
}

export interface CouponUpdate {
  isActive?: boolean;
  code?: string;
  type?: string;
  value?: number;
  minOrder?: number;
}