# lib/api-zod/src/generated/types/bannerInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BannerInput {
  id: string;
  title: string;
  subtitle?: string;
  image?: string;
  bgColor?: string;
  sortOrder?: number;
}