# lib/api-zod/src/generated/types/coupon.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Coupon {
  id: string;
  code: string;
  type: string;
  value: number;
  minOrder: number;
  isActive: boolean;
  createdAt: string;
}