# lib/api-zod/src/generated/types/menuItemInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItemInput {
  id: string;
  name: string;
  description?: string;
  price: number;
  image?: string;
  category?: string;
  rating?: number;
  isBestSeller?: boolean;
  isAvailable?: boolean;
  dailyQuota?: number;
  soldToday?: number;
  quotaResetDate?: string;
}