# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputItems } from './orderInputItems';

export interface OrderInput {
  id: string;
  items: OrderInputItems;
  customerName: string;
  whatsapp: string;
  address: string;
  notes?: string;
  paymentMethod: string;
  status?: string;
  totalPrice: number;
  subtotal: number;
  discount?: number;
  couponCode?: string | null;
}